+++ src/main.rs
@@ -2,7 +2,7 @@ use pulsar_web::{Pulse, pulsar};
 
 async fn server(mut server: Pulse) {
     server.get("/", |req, res| {
-        res.body("hello world!");
+        res.data("hello world!");
     }).await;
 
     server.launch(3000).await;